+++ database/migrations/2019_02_17_150358_create_invoices_table.php
@@ -17,7 +17,7 @@ class CreateInvoicesTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('order_id');
             $table->unsignedInteger('user_id');
-            $table->float('total_price');
+            $table->decimal('total_price',13,2);
             $table->timestamps();
 
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
