+++ database/migrations/2019_02_17_153435_create_invoice_details_table.php
@@ -18,7 +18,7 @@ class CreateInvoiceDetailsTable extends Migration
             $table->unsignedInteger('invoice_id');
             $table->unsignedInteger('product_id');
             $table->integer('quantity');
-            $table->float('price');
+            $table->decimal('price',13,2);
             $table->string('name');
             $table->string('brand');
             $table->string('category');
